EventWebsocketServer in its own file, improve err msg contexts
